Added scene serializer

+++ Atom/Source/Atom/Events/KeyEvent.h
@@ -34,7 +34,7 @@ namespace Atom
 		std::string ToString() const override
 		{
 			std::stringstream ss;
-			ss << "KeyPressedEvent: " << m_KeyCode << " (" << m_RepeatCount << " repeats)";
+			ss << "KeyPressedEvent: " << m_KeyCode << " (" << (uint32_t)m_RepeatCount << " repeats)";
 			return ss.str();
 		}
